move survey renderer to components (for easier imports in other projects)
