+++ components/survey-editor/components/ItemEditor/editor/components/content-editor/ItemPreview.tsx
@@ -1,4 +1,4 @@
-import SurveySingleItemView from '@/components/survey-viewer/survey-renderer/SurveySingleItemView/SurveySingleItemView';
+import SurveySingleItemView from '@/components/survey-renderer/SurveySingleItemView/SurveySingleItemView';
 import React, { useContext } from 'react';
 import { ComponentProperties, Expression, ExpressionArg, ItemGroupComponent, LocalizedObject, LocalizedString, SurveyItem, SurveySingleItem, isItemGroupComponent } from 'survey-engine/data_types';
 import SurveyLanguageToggle from '../../../../general/SurveyLanguageToggle';
