+++ components/survey-editor/components/ItemEditor/editor/components/content-editor/help-popup-editor.tsx
@@ -3,7 +3,7 @@ import { Switch } from '@/components/ui/switch';
 import React, { useContext } from 'react';
 import MarkdownContentEditor from './markdown-content-editor';
 import { ExpressionArg, ItemComponent, ItemGroupComponent, LocalizedString, SurveySingleItem } from 'survey-engine/data_types';
-import { getItemComponentByRole } from '@/components/survey-viewer/survey-renderer/SurveySingleItemView/utils';
+import { getItemComponentByRole } from '@/components/survey-renderer/SurveySingleItemView/utils';
 import { getLocStringLocales, localisedObjectToMap } from '@/components/survey-editor/utils/localeUtils';
 import { SurveyContext } from '@/components/survey-editor/surveyContext';
 import SurveyLanguageToggle from '@/components/survey-editor/components/general/SurveyLanguageToggle';
