+++ components/survey-renderer/SurveySingleItemView/ResponseComponent/EQ5DHealthIndicatorInput/VerticalSlider.tsx
