+++ components/survey-renderer/SurveySingleItemView/ResponseComponent/InputTypes/ClozeQuestion.tsx
