+++ components/survey-renderer/SurveySingleItemView/ResponseComponent/InputTypes/LikertGroup.tsx
