+++ components/survey-renderer/SurveySingleItemView/ResponseComponent/InputTypes/LikertScale.tsx
