+++ components/survey-renderer/SurveySingleItemView/ResponseComponent/InputTypes/MultipleChoiceGroup.tsx
