+++ components/survey-renderer/SurveySingleItemView/ResponseComponent/InputTypes/ResponsiveBipolarLikertScaleArray.tsx
