+++ components/survey-renderer/SurveySingleItemView/ResponseComponent/InputTypes/ResponsiveSingleChoiceArray.tsx
