+++ components/survey-renderer/SurveySingleItemView/ResponseComponent/InputTypes/responsiveUtils.tsx
