+++ components/survey-renderer/SurveySingleItemView/ResponseComponent/Sliders/SliderCategorical/SliderCategorical.tsx
