+++ components/survey-renderer/SurveySingleItemView/ResponseComponent/Sliders/SliderNumeric.tsx
