+++ components/survey-renderer/SurveySingleItemView/ResponseComponent/Sliders/SliderNumericRange/SliderNumericRange.tsx
