+++ components/survey-renderer/SurveySingleItemView/ResponseComponent/Sliders/SliderPrimitive.tsx
