+++ components/survey-renderer/SurveyView/SurveyProgress/SurveyProgress.tsx
