+++ components/survey-renderer/components/ConsentDialog.tsx
