+++ components/survey-renderer/components/Slider.tsx
